requirements.txt
fastapi
uvicorn
pywebview
pydantic
starlette